fix(entrypoint): steamcmd path, optional system libraries (#5)

+++ entrypoint.sh
@@ -4,9 +4,11 @@
 # Variables
 # GAME_PATH - Root game path
 # STEAMCMD_PATH - SteamCMD path
+# USE_SYS_LIBS - Use system libraries instead of the ones created by steamcmd
 #
 ##
 
+# CLI
 if [ -z ${GAME_PATH} ];
 then
     export GAME_PATH=/home/container;
@@ -14,7 +16,12 @@ fi
 
 if [ -z ${STEAMCMD_PATH} ];
 then
-    export STEAMCMD_PATH="$GAME_PATH";
+    export STEAMCMD_PATH="${GAME_PATH}/steamcmd";
+fi
+
+if [ -z ${USE_SYS_LIBS} ];
+then
+    export USE_SYS_LIBS="true"
 fi
 
 cd "$GAME_PATH"
@@ -28,17 +35,18 @@ export INTERNAL_IP=`ip route get 1 | awk '{print $NF;exit}'`
 if [ ! -z ${SRCDS_APPID} ]; then
     if [ ! -z ${SRCDS_BETAID} ]; then
         if [ ! -z ${SRCDS_BETAPASS} ]; then
-            ./$STEAMCMD_PATH/steamcmd/steamcmd.sh +login anonymous +force_install_dir "$GAME_PATH" +app_update ${SRCDS_APPID} -beta ${SRCDS_BETAID} -betapassword ${SRCDS_BETAPASS} +quit
+            $STEAMCMD_PATH/./steamcmd.sh +login anonymous +force_install_dir "$GAME_PATH" +app_update ${SRCDS_APPID} -beta ${SRCDS_BETAID} -betapassword ${SRCDS_BETAPASS} +quit
         else
-            ./$STEAMCMD_PATH/steamcmd/steamcmd.sh +login anonymous +force_install_dir "$GAME_PATH" +app_update ${SRCDS_APPID} -beta ${SRCDS_BETAID} +quit
+            $STEAMCMD_PATH/./steamcmd.sh +login anonymous +force_install_dir "$GAME_PATH" +app_update ${SRCDS_APPID} -beta ${SRCDS_BETAID} +quit
         fi
     else
-        ./$STEAMCMD_PATH/steamcmd/steamcmd.sh +login anonymous +force_install_dir "$GAME_PATH" +app_update ${SRCDS_APPID} +quit
+        $STEAMCMD_PATH/./steamcmd.sh +login anonymous +force_install_dir "$GAME_PATH" +app_update ${SRCDS_APPID} +quit
     fi
 fi
 
-# Use system libraries
-rm "${GAME_PATH}/bin/libstdc++.so.6" | true
+if [ "$USE_SYS_LIBS" = "true" ]; then
+    rm "${GAME_PATH}/bin/libstdc++.so.6" | true
+fi
 
 # Replace Startup Variables
 MODIFIED_STARTUP=`eval echo $(echo ${STARTUP} | sed -e 's/{{/${/g' -e 's/}}/}/g')`
